server/pyproject.toml
[tool.poetry]
name = "ytmusic"
version = "0.1.0"
description = ""
authors = ["Сергей Паламарчук <sergeypalamarchuk0512@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
tortoise = "^0.1.1"
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.0"}
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
aiohttp = "^3.9.5"
yt-dlp = "^2024.4.9"
setuptools = "^69.5.1"
redis = {extras = ["hiredis"], version = "^5.0.4"}
aiofiles = "^23.2.1"
pydantic = "^2.7.1"
ffmpeg-python = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
